Added MachineConfig test to the set
